LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/backward-navigation/skips-root-focusgroup-complex-case-expected.txt

item1
item3 item4
item5

FAIL When the focus is set on the last item of a focusgroup and the previous item is located past an other (non-extending) focusgroup subtree, a backward arrow key press should move the focus to that previous item without getting stuck in the other focusgroup. The same should still be true when inside a focusgroup that extends a root focusgroup within the original focusgroup. assert_equals: expected Element node <span id="item1" tabindex="0">item1</span> but got Element node <span id="item5" tabindex="-1">item5</span>
